Revert "browserify and caching update"
