+++ lib/core/environment.js
@@ -59,9 +59,6 @@ exports.getConfig = function (platform, env, cwd, argv) {
     // add Hoodie paths to config
     hoodie: {
       env: env,
-      build_debug: false,
-      from_cache: false,
-      create_date: +new Date(),
       app_path: path.resolve(cwd),
       data_path: path.resolve(path.join(cwd, 'data'))
     },
