+++ lib/helpers/plugin_api.js
