+++ lib/index.js
@@ -14,8 +14,6 @@ var server = require('./server');
 var localtld = require('./utils/localtld');
 var plugins = require('./core/plugins');
 var nodejitsu_server = require('./server/nodejitsu_server');
-var pack_hoodie = require('./utils/pack_hoodie');
-var pack_hoodie_watcher = require('./utils/pack_hoodie_watcher');
 
 var installer = require('./couchdb/installer');
 var environment = require('./core/environment');
@@ -48,17 +46,6 @@ exports.init = function (env_config, callback) {
         couch.start,
         installer.install,
         plugins.load,
-        function (config, callback) {
-          console.log('Loading hoodie.js... ');
-          var start = new Date().getTime();
-          pack_hoodie(config, function (err) {
-            var end = new Date().getTime();
-            var duration = end - start;
-            console.log('done loading hoodie.js (' + duration + 'ms)');
-            return callback(err);
-          });
-        },
-        pack_hoodie_watcher,
         hooks,
         nodejitsu_server,
         hoodieServer,
