+++ lib/server/plugins/api/index.js
@@ -2,7 +2,8 @@ var _ = require('lodash/dist/lodash.underscore');
 var path = require('path');
 var Wreck = require('wreck');
 
-var plugins = require('../../../utils/plugin_api');
+var hoodiejs = require('../../../helpers/pack_hoodie');
+var plugins = require('../../../helpers/plugin_api');
 
 var internals = {
   uiKitPath: '/node_modules/hoodie-server/node_modules/hoodie-admin-dashboard-uikit/dist/',
@@ -85,7 +86,6 @@ var internals = {
 };
 
 exports.register = function (plugin, options, next) {
-
   // FIXME: have to define these three internals here,
   //     because they rely on `options` being defined.
   //     This means we can't write unit tests for them,
@@ -108,7 +108,7 @@ exports.register = function (plugin, options, next) {
     }
   };
   internals.getHoodiePath = function (request, reply) {
-    reply(options.app.hoodiejs)
+    reply(hoodiejs(options.app))
     .type('application/javascript')
     .header(
       'Pragma', 'no-cache'
@@ -268,12 +268,6 @@ exports.register = function (plugin, options, next) {
     {
       method: 'GET',
       path: '/_api/_files/hoodie.js',
-      config: {
-        cache: {
-          privacy: 'public',
-          expiresIn: options.app.hoodie.create_date + (24 * 60 * 60 * 1000) // 1 day from create_date
-        }
-      },
       handler: internals.getHoodiePath
     }
   ]);
