+++ package.json
@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "async": "^0.9.0",
-    "browserify": "^8.1.1",
+    "browserify": "^4.2.0",
     "bytes": "^1.0.0",
     "cli-color": "^0.3.2",
     "couchr": "0.0.14",
@@ -29,7 +29,7 @@
     "request": "^2.51.0",
     "semver": "^2.3.2",
     "watch": "^0.11.0",
-    "wreck": "~5.1.0"
+    "wreck": "~5.0.1"
   },
   "scripts": {
     "start": "bin/start",
